Add info to the README, and set audio from PI as default Fixed small regression on Windows and wrap up the README
 M README.rst
Added docs link.

 M src/pupil_labs/dynamic_content_on_rim/dynamic_rim.py
 Pass int32 to circle to avoid errors (error only occurred on Windows)

+++ src/pupil_labs/dynamic_content_on_rim/dynamic_rim.py
@@ -592,7 +592,7 @@ def prepare_image(frame, xy, str, corners_screen, _screen, mheight=0, alpha=0.3)
     """
     frame = np.asarray(frame, dtype=np.float32)
     frame = frame[:, :, :]
-    xy = xy.to_numpy(dtype=np.uint64)
+    xy = xy.to_numpy(dtype=np.int32)
     # Frame to bgr
     frame = cv2.cvtColor(frame, cv2.COLOR_RGB2BGR)
     # Add screen overlay and downsize ref image
